# data-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ServerStatus 监控 API 文档",
        "title": "ServerStatus API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/export/history": {
            "get": {
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "export"
                ],
                "summary": "导出服务器历史数据为CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SessionID",
                        "name": "session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "项目密钥",
                        "name": "project_key",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "最多返回条数，默认DataLimit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "回溯小时数，默认24",
                        "name": "hours",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "缺少hostname",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/export/servers": {
            "get": {
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "export"
                ],
                "summary": "导出服务器列表为CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目密钥",
                        "name": "project_key",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "访问密钥（用于换取项目密钥）",
                        "name": "access_key",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/export/user-resources": {
            "get": {
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "export"
                ],
                "summary": "导出指定服务器的用户资源为CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主机名或SessionID",
                        "name": "hostname",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "项目密钥",
                        "name": "project_key",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "缺少hostname",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}